refactor: line station 로직 sections 객체 내부로 이동
