+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -19,8 +19,8 @@ import nextstep.subway.station.dto.StationResponse;
 @Service
 @Transactional
 public class LineService {
-    private LineRepository lineRepository;
-    private StationService stationService;
+    private final LineRepository lineRepository;
+    private final StationService stationService;
 
     public LineService(LineRepository lineRepository, StationService stationService) {
         this.lineRepository = lineRepository;
@@ -33,7 +33,7 @@ public class LineService {
         Line persistLine = lineRepository.save(
             new Line(request.getName(), request.getColor(), upStation, downStation, request.getDistance()));
         List<StationResponse> stations = persistLine.getStations().stream()
-            .map(it -> StationResponse.of(it))
+            .map(StationResponse::of)
             .collect(Collectors.toList());
         return LineResponse.of(persistLine, stations);
     }
@@ -43,7 +43,7 @@ public class LineService {
         return persistLines.stream()
             .map(line -> {
                 List<StationResponse> stations = line.getStations().stream()
-                    .map(it -> StationResponse.of(it))
+                    .map(StationResponse::of)
                     .collect(Collectors.toList());
                 return LineResponse.of(line, stations);
             })
@@ -57,7 +57,7 @@ public class LineService {
     public LineResponse findLineResponseById(Long id) {
         Line persistLine = findLineById(id);
         List<StationResponse> stations = persistLine.getStations().stream()
-            .map(it -> StationResponse.of(it))
+            .map(StationResponse::of)
             .collect(Collectors.toList());
         return LineResponse.of(persistLine, stations);
     }
@@ -75,7 +75,7 @@ public class LineService {
         Line line = findLineById(lineId);
         Station upStation = stationService.findStationById(request.getUpStationId());
         Station downStation = stationService.findStationById(request.getDownStationId());
-        line.addSection(new Section(line, upStation, downStation, request.getDistance()));
+        line.addLineStation(new Section(line, upStation, downStation, request.getDistance()));
     }
 
     public void removeLineStation(Long lineId, Long stationId) {
